Add Ticke seeding
