+++ src/backend/database/seeders/TicketSeeder.php
@@ -7,8 +7,15 @@ use App\Models\GeographicalInformation;
 use App\Models\MaintenanceUsers;
 use App\Models\MiniGrid;
 use App\Models\Person\Person;
+use App\Models\User;
+use Illuminate\Console\View\Components\Info;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
+use Inensus\Ticket\Models\Ticket;
 use Inensus\Ticket\Models\TicketCategory;
+use Inensus\Ticket\Models\TicketOutsource;
+use Inensus\Ticket\Models\TicketUser;
 use MPM\DatabaseProxy\DatabaseProxyManagerService;
 
 class TicketSeeder extends Seeder {
@@ -18,12 +25,18 @@ class TicketSeeder extends Seeder {
         $this->databaseProxyManagerService->buildDatabaseConnectionDummyCompany();
     }
 
+    private $amount = 100;
+
     /**
      * Run the database seeds.
      *
      * @return void
      */
     public function run() {
+        (new Info($this->command->getOutput()))->render(
+            "Running TransactionSeeder to generate $this->amount tickets. This may take some time."
+        );
+
         // Create Ticket categories
         TicketCategory::factory()
             ->count(12)
@@ -130,5 +143,83 @@ class TicketSeeder extends Seeder {
                 ->for($agentPerson)
                 ->create();
         }
+
+        // Seed tickets
+        for ($i = 1; $i <= $this->amount; ++$i) {
+            try {
+                DB::connection('shard')->beginTransaction();
+                $this->generateTicket();
+                DB::connection('shard')->commit();
+            } catch (\Exception $e) {
+                DB::connection('shard')->rollBack();
+                echo $e->getMessage();
+            }
+        }
+    }
+
+    private function generateTicket() {
+        $randomCategory = TicketCategory::query()->inRandomOrder()->first();
+        $fakeSentence = $this->generateFakeSentence();
+        $randomCreator = User::inRandomOrder()->first();
+        $demoDate = date('Y-m-d', strtotime('-'.mt_rand(0, 365).' days'));
+        $ticketUser = TicketUser::inRandomOrder()->first();
+        $randomMaintenanceUser = MaintenanceUsers::inRandomOrder()->first();
+        $randomPerson = Person::inRandomOrder()->where('is_customer', 1)->first();
+        $dueDate = date('Y-m-d', strtotime('+3 days', strtotime($demoDate)));
+        $status = rand(0, 1);
+
+        $ticket = Ticket::query()->make([
+            'ticket_id' => Str::random(10),
+            'creator_type' => 'admin',
+            'creator_id' => $randomCreator->id,
+            'status' => $status,
+            'due_date' => $dueDate,
+            'title' => 'Dummy Ticket',
+            'content' => $fakeSentence,
+            'category_id' => $randomCategory->id,
+            'created_at' => $demoDate,
+            'updated_at' => $demoDate,
+        ]);
+
+        if ($randomCategory->out_source) {
+            $ticket->assigned_id = null;
+            $ticket->owner_id = $randomMaintenanceUser->id;
+            $ticket->owner_type = 'maintenance_user';
+            $ticket->save();
+            try {
+                $amount = random_int(10, 200);
+            } catch (\Exception $e) {
+                $amount = 50;
+            }
+            TicketOutsource::query()->create([
+                'ticket_id' => $ticket->id,
+                'amount' => $amount,
+                'created_at' => $demoDate,
+                'updated_at' => $demoDate,
+            ]);
+        } else {
+            $ticket->assigned_id = $ticketUser->id;
+            $ticket->owner_id = $randomPerson->id;
+            $ticket->owner_type = 'person';
+            $ticket->save();
+        }
+    }
+
+    private function generateFakeSentence($minWords = 5, $maxWords = 15) {
+        $loremIpsum =
+            'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.';
+        $words = explode(' ', $loremIpsum);
+        $numWords = rand($minWords, $maxWords);
+
+        shuffle($words);
+        $fakeSentence = implode(' ', array_slice($words, 0, $numWords));
+
+        // Capitalize the first letter of the sentence.
+        $fakeSentence = ucfirst($fakeSentence);
+
+        // Add a period at the end.
+        $fakeSentence .= '.';
+
+        return $fakeSentence;
     }
 }
